+++ docs/api/syntax.rst
@@ -13,6 +13,7 @@ DisCoPy's mathematical core: a zoo of diagrams, categories and functors.
     discopy.balanced
     discopy.symmetric
     discopy.markov
+    discopy.closed
     discopy.feedback
     discopy.traced
     discopy.biclosed
